jwt authentication

+++ backend/controllers/goalController.js
@@ -1,11 +1,13 @@
 const asyncHandler = require('express-async-handler')
+
 const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
+
 // @desc    Get Goals
 // @route   GET /api/goals
 // @access  Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find()
-
+  const goals = await Goal.find({ user: req.user.id })
   res.status(200).json(goals)
 })
 
@@ -18,7 +20,7 @@ const setGoal = asyncHandler(async (req, res) => {
     throw new Error('Please add a text field')
   }
 
-  const goal = await Goal.create({ text: req.body.text })
+  const goal = await Goal.create({ text: req.body.text, user: req.user.id })
 
   res.status(201).json(goal)
 })
@@ -32,6 +34,18 @@ const updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400)
     throw new Error('Goal not found')
+  }
+
+  // Check for user
+  if (!req.user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -49,6 +63,18 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400)
     throw new Error('Goal not found')
+  }
+
+  // Check for user
+  if (!req.user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  // Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== req.user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
   }
 
   await goal.deleteOne()
